Spawn separate system process for PhantomJS
